requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
PyPDF2
python-multipart
requests
beautifulsoup4
playwright
duckdb
google-generativeai
